+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `page-meta` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0 - 2026-04-10
 
-- initial release
+- planned: HtmlMeta support (title, description, author, canonical, robots/noindex)
+
+## 0.1.2 - 2023-01-22
+
+- Add Twitter Cards
+
+## 0.1.1 - 2023-01-22
+
+- Fix minor issues
+
+## 0.1.0 - 2023-01-08
+
+- Initial release: OpenGraph meta tag support